+++ cpp/dolfinx/graph/partition.cpp
@@ -8,6 +8,8 @@
 #include "partitioners.h"
 #include <algorithm>
 #include <dolfinx/common/IndexMap.h>
+#include <dolfinx/common/MPI.h>
+#include <dolfinx/common/Timer.h>
 #include <dolfinx/common/log.h>
 #include <dolfinx/graph/AdjacencyList.h>
 #include <memory>
